moved contents stock_tracker_app around for organization
random object and functions taken to extras
create monitor file to hold only monitor material
account_manage for login related material
